requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
defusedxml==0.7.1
Django==5.0.1
django-bootstrap4==24.1
django-crispy-forms==2.1
django-cron==0.6.0
djangorestframework==3.14.0
fontawesomefree==6.5.1
httpie==3.2.2
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
pillow==10.2.0
psycopg==3.1.18
Pygments==2.17.2
PySocks==1.7.1
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.0
